Optional feedback, timer and counter
This is my first ever pull request so please let me know if I should do anything differently!

I would like to make several updates to the memory game to make it suitable for young children that cannot read. This first pull request is aimed at making the feedback, timer and counter html fields optional.

The way I've done this is to set the containing elements to hidden when initialized, and thereafter display them if this is set in the behavior settings.